Change default file handling (RhBug: 1094539)
During update, the original createrepo leaves behind some files which
were originaly part of repodata (were listed in the repomd.xml)
and after the update they are not part of the repomd.xml but they
still exists in the repodata/ directory.

E.g. comps files (group files) - when you add a new group file
(using --groupfile) and use --update, the old groupfile remains
are in the repodata/.

This patch changes the approach and the files that are not
part of repomd.xml anymore are removed.

See: https://github.com/Tojaj/createrepo_c/wiki/New-File-Handling