Fix double free during parsing broken XML metadata (Issue #33)
The issue was reported by Rohan McGovern including excelent
reproducer and detail problem analysis.

Rohan's theory that was absolutely right:

When running in single chunk mode, there is one GStringChunk
which belongs to the cr_Metadata. It's temporarily
assigned/unassigned to/from cr_Package objects in newpkgcb/pkgcb
in load_metadata.c.

If XML parsing stops in the middle of a package element,
pkgcb is not called for the last cr_Package, so the chunk
remains assigned to that object.

The chunk is then freed twice, once when the hash table of
packages is destroyed (end of cr_metadata_load_xml function)
and again when the cr_Metadata object is destroyed.