Misc: Add cr_version_string_with_features() function
This function returns a static string which contains
version of createrepo_c library and list of features
which were enabled during compilation time.

Currently there are only two features:
"DeltaRPM" which means that createrepo_c lib was compiled
with support for generation of Delta RPMS.
"LegacyWeakdeps" which means that createrepo_c was compiled
with support for legacy weakdeps which are (were?) used in
OpenSuse and related distros.

Example string:
"0.10.0 (Features: DeltaRPM )"

Signed-off-by: Tomas Mlcoch <tmlcoch@redhat.com>