Don't replace metadata on --update if we've got 100% cache hits
In the case where createrepo_c has been called with "--update", if the
repodata for every package we find is satisfied from the existing cache,
do not update the metadata files at all.

This means rsync doesn't have to evaluate the files for changes, and in
the case where '--unique-md-filenames --update' is used, it means rsync
won't have to transfer the newly renamed files.  This saves
approximately 8k of bandwidth (and many round trips) per rsync run in
the cases where nothing in the repo has changed, and also reduces the
time during a scheduled batch job where the repo is out of sync.

In the '--unique-md-filenames --update' case, it also means it won't
rename the files (actually representing only a newer generation of the
exact same data) if '--simple-md-filenames' was previously used on the
same repository.

Signed-off-by: Peter Jones <pjones@redhat.com>