Add libglnx GLib cleanup backports, use in a new place
Now that the cleanup macros are upstream in GLib, let's follow
the path of e.g. ostree/rpm-ostree and many other projects that
use them.  I added the "backports" bits from libglnx:
https://git.gnome.org/browse/libglnx

I left in place the previous `cleanup.h`, but at some point we
should try to do a mass conversion.

To demonstrate use, I changed one place, but there is *tons* of code here that
could be cleaned up. Note also the use of "move semantics" there.

I'm writing this patch as it's likely I might start some work
on this project as part of the rpm-ostree jigdo ♲📦 effort:
https://github.com/projectatomic/rpm-ostree/issues/1081