Add --pkgorigins mode
When handling modules, we want a mode of operation for Koji that
handles the pkgorigins file but does not do any of the other
filtering or blocking that the --koji mode does.

Note: this patch doesn't need to change any code where the
pkgorigins file is actually populated, because it was already
conditionalized merely on the koji_stuff->pkgorigins pointer
being non-NULL, not driven by the cmd_options->koji.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>