[spec] Obsolete createrepo on all Fedoras again (RhBug:1702771)
When upgrading createrepo (legacy) on a Fedora 30 system, DNF replaces
(obsoletes) it with createrepo_c.  However, with the conditionals added
in commit 2090473b, DNF no longer picks the latest available version of
createrepo_c from the updates repo.  Instead, the pre-2090473b version
is selected from the stable (frozen) repo.

To fix that, this commit partially reverts the above commit so that we
obsolete everywhere but on RHEL-7 where createrepo is still a core
component with full support.

Reproducer:
$ dnf install createrepo.noarch  # installs the legacy createrepo
$ dnf upgrade createrepo

Before: Upgrades to createrepo_c from stable
After:  Upgrades to createrepo_c from updates