Fix null pointer crash in cr_dumper_thread() if it fails to get checksum
If the call to cr_checksum_file() inside get_checksum() fails, tmp_err
is propagated to err and makes it back to the caller load_rpm(). But
load_rpm() doesn't propagate the error back to its caller
cr_dumper_thread(). As a result when cr_dumper_thread() tries to print a
warning, it crashes with a null pointer exception on tmp_err.

Signed-off-by: Gaurav Gupta <gauragup@cisco.com>