Add a new function to replace PyObject_ToStrOrNull()
There's no good way to make PyObject_ToStrOrNull() work. It requires
allocating an intermediate value, and we can't deallocate it because
the return value points into the buffer.

Need to just return the PyBytes PyObject so that it's lifetime can be
managed by the caller. Additionally, the PyObject_ToChunkedString()
function seems under-utilized, we can improve it and use it to keep the
ergonomics that we had before.