Expand documentation, add warning
Help text:

```
$ createrepo_c --help-expert
Usage:
  createrepo_c [OPTION?] <directory_to_index>

Program that creates a repomd (xml-based rpm metadata) repository from a set of rpms.

Expert (risky) options
  --ignore-lock                                Expert (risky) option:
Ignore an existing .repodata/. (Remove the existing .repodata/ and create an empty new one to serve as a lock for other createrepo instances. For the repodata generation, a different temporary dir with the name in format ".repodata.time.microseconds.pid/" will be used).  NOTE: Use this option on your own risk! If two createrepos run simultaneously, then the state of the generated metadata is not guaranteed - it can be inconsistent and wrong.
  --skip-filelists                             Expert (risky) option:
Skip filelist generation, potentially saving significant bandwidth for repos with large numbers of files in packages. NOTE: Use this option on your own risk! This is a site-local optimization and should not be used for public repos. The site operator warrants the filenames in the packages in the repo are not named as required in other packages. The site operator also warrants that all clients do not need to use repoquery -l and -f to list or find packages that own a given file.
```

Example run:

```
$ createrepo_c --skip-filelists .
Directory walk started
Directory walk done - 9 packages
Temporary output repo path: ./.repodata/
Preparing sqlite DBs
Warning: Expert option: --skip-filelists for site-local optimization active
Pool started (with 5 workers)
Pool finished
```