Wrap c api to python for parsing metadata together
It is not sufficient to store a reference to a single currently
processed package because xml_parse_main_metadata_together can have
multiple packages in progress at the same time.

This is caused by the way parsing is done, since we are reading metadata
in larger sections and we don't know if a given section has just one
pkg, half of a pkg or several pkgs (they have varying lenghts).
It is possible we read part of a package in primary and then several
packages in filelists. This causes the newpkgcb to be called several
times in a row before a single package is finished.

We also need the pkg dictionary if pkgs are out of order among the 3
metadata files, although this will consume more and more memory.