If new and old repomd matches during --update don't update
If the repomd matches it contains the same data. By not touching the
repodata files we can potentially save resources because other tools
don't have to evaluate the files for changes.

= changelog =
msg: If the new repodata generated during an --update run exactly matches the old repodata don't touch the files.
type: enhancement