Free `in_progress_pkgs_list` fully only when there is no newpkgcb
Normally the `in_progress_pkgs_list` is empty (all packages get parsed
and processed) however there are cases such as hitting an error or
interrupting parsing when there can be packages present.

If there are packages we can free them only if there is no newpkgcb
because otherwise their references are stored in python layer which then
also tries to free them leading to a crash.