Update errno usage to fix incorrect GError messages
Either the call to g_strerror or the GLib logging calls have been
changing the value of `errno` before it is used to set the error message
in g_set_error. This resulted in a correct error message being logged,
but an incorrect message being reported to the caller.

This change resolves the issue by calling g_strerror(errno) once and
re-using the same message to pass to both the logging function and
g_set_error.