New (hidden) --delayed-dump option
Normally we dump (to XML and DB) directly from the pool of threading
workers, in a streaming fashion (as soon as the metadata about packages
are determined it is written).

The new option causes that the metadata are loaded first (still in
parallel), and then written at one time (by one thread for now).
Since all the metadata (cr_Package) is available in the memory in one
time, we can do some additional processing and filtering (before we
eventually perform the dump).

This is an optional behavior because the delayed XML chunk preparation
slows down the overall multi-threaded processing time.