Make the --update old_metadata override hack simpler
Even though we are destructively changing the recycled cr_Packages data,
we can keep the structure self-contained so the upper level logic will
automatically take care of the allocated memory.

The recycled metadata cr_Packages use one chunk per all packages
(CR_PACKAGE_SINGLE_CHUNK).  That's why (pkg->chunk == NULL) and we can
re-use it for a new chunk.  We just need to change the type of pkg to
!CR_PACKAGE_SINGLE_CHUNK so the cr_package_free() cleans the new
chunk.

This allows us to stop taking care of the locally allocated location_*
variables, because they are _cleanup_free_.  This makes even the buffer
handling simpler.