Avoid drpm regeneration
If a drpm file for the same target already exists, we don't
need to process the original files again.

We can't check the source rpm only the target rpm hashsum, but as drpm
is content based not rpm based this should be could enough.

If the target is overridden with different content but has the same
name, we need to recalculate the delta rpm.