Detect need for _XOPEN_SOURCE for nftw() in CMakeLists.txt
Some systems (e.g., glibc Linux systems) required _XOPEN_SOURCE be set
to >= 500 in order to expose the nftw() function.  Other systems, such
as macOS, do not require this definition.  When using feature macros
like this, it is better to not define them explicitly in source code
but rather define them dynamically at build configuration time based
on what the build system presents.

This patch removes the explicit definiton of _XOPEN_SOURCE from
src/misc.c and adds a block to src/CMakeLists.txt to determine if
_XOPEN_SOURCE must be defined in order to use nftw().  If it is
required, I set it to 700 which is what is required for the Python
code as well and it avoids a redefinition warning.

Fixes: #374

Signed-off-by: David Cantrell <dcantrell@redhat.com>