Default changelog limit restored for --compatibility
Defaulting to keeping all changelogs, while accurate to the behavior of
legacy createrepo, can result in terrible behaviors for some
repositories, making them potentially too large to be usable (libsolv
may segfault, I'm not sure about YUM, but both end up unpacking the
data on the root partition potentially causing it to run out of space).

This behavior of the --compatibility flag was never explicitly documented.

closes #318