Use IMPORTED_TARGET for 3rd-party dependencies
The current CMakeLists.txt fails to include the required link directories for 3rd-party packages.
As an example, on FreeBSD where packages are installed under /usr/local, the link lines include -lgio-2.0 but not -L/usr/local/lib.

The suggested solution is to use the IMPORTED_TARGET mode of pkg_check_modules. This requires CMake 3.6, so I have bumped the minimum required version.